4-new_dog.c 2

+++ 0x0E-structures_typedef/4-new_dog.c
@@ -9,9 +9,9 @@
  * Return: The length of the string
  */
 
-size_t _strlen(const char *str)
+int _strlen(const char *str)
 {
-	size_t length = 0;
+	int length = 0;
 
 	while (str[length] != '\0')
 	{
